Added support for updating bounding boxes and polygons
Loading previously saved records will now display the appropriate bounding box or polygon on the map, if either exists, via the whenReady map event.

Using the delete button will also clear the bounding box text fields as well as the polygon field.

Max bounds have been set on the map to prevent strange coordinates from being generated by drifting too far east or west when dragging the map

Code to remove extra layers has been extracted to its own function and will now run on the number of existing layers rather than the index - this was causing issues because of the conditional rendering of bounding boxes & polygons, if you stuck with one or the other it would be fine, but if you switched from bounding box to polygon (or vice versa) it would lead to multiple shapes accruing.

functions have been created to test for valid bounding box coordinates and polygon points to prevent crashing.

Updating bounding box text fields or the polygon field will result in the updating of the current shape (and removing other shapes)

Updated height of map to use 55% of viewport height.