Added validation counts for optional fields in addition to required.
This allows optional fields with validation to be appropriately flagged without requiring a required field to be in error first.

However, this fix is somewhat redundant since it effectively is checking all fields in two separate groups and adding them together.  Fields without validation get a pass by default which means there's no penalty to simply check all fields and let the validator run normally.

The only question is if there's any value in splitting out required and optional fields at this stage.